+++ Assets/Hotfix/Runtime/ECS/Component/ComOwnerEntity.cs
@@ -2,7 +2,7 @@ namespace LccHotfix
 {
     public class ComOwnerEntity : LogicComponent
     {
-        public int ownerEntityID;
+        public long ownerEntityID;
     }
 
     public partial class LogicEntity
@@ -10,7 +10,7 @@ namespace LccHotfix
         public ComOwnerEntity comOwnerEntity { get { return (ComOwnerEntity)GetComponent(LogicComponentsLookup.ComOwnerEntity); } }
         public bool hasComOwnerEntity { get { return HasComponent(LogicComponentsLookup.ComOwnerEntity); } }
 
-        public void AddComOwnerEntity(int newOwnerEntityID)
+        public void AddComOwnerEntity(long newOwnerEntityID)
         {
             var index = LogicComponentsLookup.ComOwnerEntity;
             var component = (ComOwnerEntity)CreateComponent(index, typeof(ComOwnerEntity));
@@ -18,7 +18,7 @@ namespace LccHotfix
             AddComponent(index, component);
         }
 
-        public void ReplaceComOwnerEntity(int newOwnerEntityID)
+        public void ReplaceComOwnerEntity(long newOwnerEntityID)
         {
             var index = LogicComponentsLookup.ComOwnerEntity;
             var component = (ComOwnerEntity)CreateComponent(index, typeof(ComOwnerEntity));
